requirements.txt
matplotlib
numpy
pandas
Pillow
pylsl
PyQt5
pyqtgraph
scikit-learn
scipy
six
medusa-kernel